Moving parameters into environment variables
