+++ src/ros2/launch_application.sh
@@ -1,2 +1,5 @@
 #!/bin/bash
+export BOB_SOURCE="'video'"
+export BOB_ENABLE_VISUALISER="True"
+
 ./launcher.sh application_launch.py
