+++ src/ros2/src/bob_launch/launch/application_launch.py
@@ -1,45 +1,51 @@
 import os
 from launch import LaunchDescription
-from launch.actions import IncludeLaunchDescription
-from ament_index_python.packages import get_package_share_directory
+from launch.actions import IncludeLaunchDescription, DeclareLaunchArgument
+from launch.substitutions import EnvironmentVariable
 from launch.launch_description_sources import PythonLaunchDescriptionSource, FrontendLaunchDescriptionSource
+from ament_index_python.packages import get_package_share_directory
 
 def generate_launch_description():
     
+    #https://roboticscasual.com/tutorial-ros2-launch-files-all-you-need-to-know/
+
     launch_package_dir = get_package_share_directory('bob_launch')
     ros_bridge_package_dir = get_package_share_directory('rosbridge_server')
 
+    source_arg_value = EnvironmentVariable('BOB_SOURCE', default_value="'video'") 
+    enable_visualiser_arg_value = EnvironmentVariable('BOB_ENABLE_VISUALISER', default_value="True")
+    enable_rosbridge_arg_value = EnvironmentVariable('BOB_ENABLE_ROSBRIDGE', default_value="False")
+
+    source_arg = DeclareLaunchArgument(
+        'source_arg',
+        default_value=source_arg_value,
+        description="Argument for the image source of the application."
+        )
+
+    enable_visualiser_arg = DeclareLaunchArgument(
+        'enable_visualiser_arg',
+        default_value=enable_visualiser_arg_value,
+        description="Argument for the enabling of the visualiser screen."
+        )
+
+    enable_rosbridge_arg = DeclareLaunchArgument(
+        'enable_rosbridge_arg',
+        default_value=enable_rosbridge_arg_value,
+        description="Argument for the enabling of the ROS Bridge."
+        )
+
     return LaunchDescription([
 
-        #
-        # Uncomment this IncludeLaunchDescription section if you would like to launch using the supplied videos as the image source
-        # If you are using a different source, please exclude this IncludeLaunchDescription by commenting it out        
-        #
+        source_arg,
+        enable_visualiser_arg,
+        enable_rosbridge_arg,
+        
+
         IncludeLaunchDescription(
             PythonLaunchDescriptionSource([
                 launch_package_dir, 
-                '/video_kernel_launch.py'])
+                '/kernel_launch.py']),
         ),
-
-        #
-        # Uncomment this IncludeLaunchDescription section if you would like to launch using an RTSP camera as the image source
-        # If you are using a different source, please exclude this IncludeLaunchDescription by commenting it out
-        #
-        #IncludeLaunchDescription(
-        #    PythonLaunchDescriptionSource([
-        #        launch_package_dir, 
-        #        '/rtsp_kernel_launch.py'])
-        #),        
-
-        #
-        # Uncomment this IncludeLaunchDescription section if you would like to launch using a USB camera as the image source
-        # If you are using a different source, please exclude this IncludeLaunchDescription by commenting it out
-        #
-        #IncludeLaunchDescription(
-        #    PythonLaunchDescriptionSource([
-        #        launch_package_dir, 
-        #        '/usb_kernel_launch.py'])
-        #),  
 
         IncludeLaunchDescription(
             PythonLaunchDescriptionSource([
@@ -55,6 +61,9 @@ def generate_launch_description():
 
         IncludeLaunchDescription(
             FrontendLaunchDescriptionSource(
-                os.path.join(ros_bridge_package_dir,
-                             'launch/rosbridge_websocket_launch.xml')))
+                os.path.join(
+                    ros_bridge_package_dir,
+                    'launch/rosbridge_websocket_launch.xml'))
+        )
+
     ])
