+++ src/ros2/src/bob_launch/launch/display_launch.py
@@ -1,6 +1,9 @@
 from launch import LaunchDescription
 from launch_ros.actions import ComposableNodeContainer
 from launch_ros.descriptions import ComposableNode
+from launch.actions import LogInfo
+from launch.substitutions import PythonExpression, LaunchConfiguration
+from launch.conditions import IfCondition
 
 def generate_launch_description():
 
@@ -15,7 +18,8 @@ def generate_launch_description():
                 plugin='FrameViewer',
                 name='frame_viewer_node',
                 parameters=[{"topics": ["bob/camera/all_sky/bayer/resized", "bob/frames/all_sky/foreground_mask/resized", "bob/frames/annotated/resized"]}],
-                extra_arguments=[{'use_intra_process_comms': True}]),
+                extra_arguments=[{'use_intra_process_comms': True}],
+                condition=IfCondition(PythonExpression([LaunchConfiguration('enable_visualiser_arg'), " == True" ])),),
             ],
             output='screen',
     )
@@ -52,4 +56,13 @@ def generate_launch_description():
         output='screen',
     )
 
-    return LaunchDescription([display_container, compress_container])
+    return LaunchDescription([
+
+        LogInfo(
+            condition=IfCondition(PythonExpression([LaunchConfiguration('enable_visualiser_arg'), " == False" ])),
+            msg=['Visualiser not active.']
+            ),
+
+        display_container, 
+        compress_container
+        ])
